reset invalid profile (#373)
* change button name (fix #341)
  * invalid profile was removed from the login list (fixes #353)
  * the profile can be reset after an unexpected exit

+++ src/views/pages/profiles/import-profile/import-mnemonic/ImportMnemonicTs.ts
@@ -73,10 +73,11 @@ export default class ImportMnemonicTs extends Vue {
    * @return {void}
    */
   public deleteProfileAndBack() {
-    // - delete the temporary profile from storage
-    this.profileService.deleteProfile(this.currentProfile.profileName)
-    this.$store.dispatch('profile/RESET_STATE')
-
+    if (this.currentProfile) {
+      // - delete the temporary profile from storage
+      this.profileService.deleteProfile(this.currentProfile.profileName)
+      this.$store.dispatch('profile/RESET_STATE')
+    }
     // - back to previous page
     this.$router.push({ name: 'profiles.importProfile.info' })
   }
